# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CareLink API
  version: 0.0.1
  description: >-
    Caregiver slice 1 is implemented for GET /caregivers/me, GET /caregivers/me/schedule,
    and GET /visits/{visitId}/work-pack; these definitions supersede the matching draft operations.
    Includes the existing Auth contract and the specified FM01-FM05 family contracts. Family
    operations remain x-status: draft until implemented; inclusion in Final does not imply a working
    backend. Shared paths retain their non-family response alternatives. The family owner supplies
    only the family use-case behaviour; other use cases, binding management and platform
    implementation are outside this change.
  license:
    name: UNLICENSED
servers:
  - url: /api
    description: Same-origin, relative to whichever host serves the front end.
...
paths:
...
  /absences:
    get:
      tags:
        - Absences
      summary: List absences with how far their re-rostering has got
      operationId: listAbsences
      x-status: implemented
      x-use-cases:
        - UC-MG04
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
      responses:
        '200':
          description: Absences, latest start first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AbsenceSummary'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - sessionCookie: []
    post:
      tags:
        - Absences
      summary: Record a caregiver's absence (UC-MG04 step 1)
      description: >-
        A manager records an absence they were told about, e.g. a caregiver ringing in sick. It is
        approved at once, because the manager is the reviewer. 409 ABSENCE_OVERLAPS when it shares a
        day with an absence already on record that was not rejected.
      operationId: recordAbsence
      x-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordAbsenceRequest'
      responses:
        '201':
          description: The absence, approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
        - sessionCookie: []
  /absences/{id}:
    parameters:
      - $ref: '#/components/parameters/AbsenceId'
    get:
      tags:
        - Absences
      summary: One absence, its unhandled visits and every change with its candidates
      description: >-
        For each vacated visit: where its change stands, and every caregiver the latest search
        considered, suggested or excluded, with every rule result behind them - the answer to "why
        was she not suggested".
      operationId: getAbsence
      x-status: implemented
      responses:
        '200':
          description: The absence in full.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceCase'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - sessionCookie: []
  /absences/{id}/approve:
    parameters:
      - $ref: '#/components/parameters/AbsenceId'
    post:
      tags:
        - Absences
      summary: Approve an absence a caregiver asked for
      operationId: approveAbsence
      x-status: implemented
      responses:
        '200':
          description: The absence, approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceSummary'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
        - sessionCookie: []
  /absences/{id}/reject:
    parameters:
      - $ref: '#/components/parameters/AbsenceId'
    post:
      tags:
        - Absences
      summary: Reject an absence a caregiver asked for
      operationId: rejectAbsence
      x-status: implemented
      responses:
        '200':
          description: The absence, rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceSummary'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
        - sessionCookie: []
  /absences/{id}/rerostering-runs:
    parameters:
      - $ref: '#/components/parameters/AbsenceId'
    post:
      tags:
        - Absences
      summary: Re-roster now (UC-MG04 steps 2-4)
      description: >-
        Searches every visit the absence vacates that has no change yet, ranked by the chosen
        objective. A visit somebody can take is offered to the family, or settled by the default plan
        at once when it starts too soon to ask; a visit nobody can take becomes an exception with an
        incident. Visits left uncovered earlier are searched again and taken straight away when
        somebody has become free. Safe to repeat. 409 OBJECTIVE_NOT_SUPPORTED for COST, 409
        ABSENCE_NOT_APPROVED for an absence still waiting for review.
      operationId: rerosterAbsence
      x-status: implemented
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                objective:
                  type: string
                  enum: [CONTINUITY, EVEN_WORKLOAD, TRAVEL_TIME]
                  default: CONTINUITY
      responses:
        '201':
          description: What the run did, and the absence as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReRostered'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
        - sessionCookie: []
  /absences/{id}/coverage-confirmation:
    parameters:
      - $ref: '#/components/parameters/AbsenceId'
    post:
      tags:
        - Absences
      summary: Confirm every vacated visit is accounted for (UC-MG04 step 7)
      description: >-
        409 FAMILY_STILL_DECIDING while a family is still answering, 409 VISITS_NOT_REROSTERED while
        vacated visits have not been searched.
      operationId: confirmAbsenceCoverage
      x-status: implemented
      responses:
        '200':
          description: The absence in full, with the confirmation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceCase'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
        - sessionCookie: []
  /caregivers/me/absences:
    get:
      tags:
        - Absences
      summary: The signed-in caregiver's own absences (UC-CG02)
      operationId: listOwnAbsences
      x-status: implemented
      responses:
        '200':
          description: Own absences, latest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OwnAbsence'
      security:
        - sessionCookie: []
    post:
      tags:
        - Absences
      summary: Ask for leave (UC-CG02); a manager reviews it
      operationId: requestAbsence
      x-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestAbsenceRequest'
      responses:
        '201':
          description: The request, waiting for review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnAbsence'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
        - sessionCookie: []
...
components:
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string

    # Family projections and referenced shared schemas.
...
    AbsenceSummary:
      type: object
      description: An absence and how far its re-rostering has got (UC-MG04).
      properties:
        id: {type: integer, format: int64}
        caregiverId: {type: integer, format: int64}
        caregiverName: {type: string}
        type: {type: string, enum: [SICK, ANNUAL, EMERGENCY, OTHER]}
        startDate: {type: string, format: date}
        endDate: {type: string, format: date}
        reason: {type: string, nullable: true}
        status: {type: string, enum: [PENDING, APPROVED, REJECTED]}
        reviewedByUserId: {type: integer, format: int64, nullable: true}
        coverageConfirmedAt: {type: string, format: date-time, nullable: true}
        notYetRerostered: {type: integer, description: Vacated visits nobody has searched for yet.}
        awaitingFamily: {type: integer}
        uncovered: {type: integer}
        settled: {type: integer}
    VacatedVisit:
      type: object
      properties:
        visitId: {type: integer, format: int64}
        elderId: {type: integer, format: int64}
        elderName: {type: string}
        serviceType: {type: string, nullable: true}
        start: {type: string, format: date-time}
        end: {type: string, format: date-time, nullable: true}
    RosterPerson:
      type: object
      nullable: true
      properties:
        caregiverId: {type: integer, format: int64}
        name: {type: string}
    RosterCheck:
      type: object
      description: One rule's result for one candidate.
      properties:
        code: {type: string, example: CERTIFICATION_VALID}
        name: {type: string}
        kind: {type: string, enum: [HARD, SOFT]}
        result: {type: string, enum: [PASS, FAIL, NOT_APPLICABLE]}
        detail: {type: string, nullable: true, example: Nursing valid to 2027-01-01}
    RosterCandidate:
      type: object
      properties:
        caregiverId: {type: integer, format: int64}
        name: {type: string}
        rank: {type: integer, nullable: true, description: 1 for the best suggestion; null when excluded.}
        score: {type: number, nullable: true}
        reason: {type: string, nullable: true, description: 'The match reason, or why a hard rule excluded them.'}
        outcome: {type: string, enum: [SELECTED, SUGGESTED, EXCLUDED]}
        excludedBy: {type: string, nullable: true}
        checks:
          type: array
          items:
            $ref: '#/components/schemas/RosterCheck'
    RosterChangeView:
      type: object
      description: One vacated visit as the manager sees it.
      properties:
        id: {type: integer, format: int64}
        visitId: {type: integer, format: int64}
        elderId: {type: integer, format: int64}
        elderName: {type: string}
        visitStart: {type: string, format: date-time}
        visitEnd: {type: string, format: date-time, nullable: true}
        status: {type: string, enum: [AWAITING_FAMILY, UNCOVERED, RESOLVED]}
        outcome: {type: string, nullable: true, enum: [REPLACED, RESCHEDULED, SKIPPED, WITHDRAWN]}
        decidedBy: {type: string, nullable: true, enum: [FAMILY, DEFAULT_PLAN, MANAGER]}
        decidedAt: {type: string, format: date-time, nullable: true}
        respondBy: {type: string, format: date-time, nullable: true}
        note: {type: string, nullable: true}
        absentCaregiver: {$ref: '#/components/schemas/RosterPerson'}
        proposedCaregiver: {$ref: '#/components/schemas/RosterPerson'}
        assignedCaregiver: {$ref: '#/components/schemas/RosterPerson'}
        rescheduledVisitId: {type: integer, format: int64, nullable: true}
        rescheduledStart: {type: string, format: date-time, nullable: true}
        incidentId: {type: integer, format: int64, nullable: true}
        rosteringRunId: {type: integer, format: int64, nullable: true}
        objective: {type: string, nullable: true, enum: [CONTINUITY, EVEN_WORKLOAD, TRAVEL_TIME]}
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/RosterCandidate'
    AbsenceCase:
      type: object
      properties:
        absence: {$ref: '#/components/schemas/AbsenceSummary'}
        notYetRerostered:
          type: array
          items:
            $ref: '#/components/schemas/VacatedVisit'
        changes:
          type: array
          items:
            $ref: '#/components/schemas/RosterChangeView'
    ReRostered:
      type: object
      properties:
        outcome:
          type: object
          properties:
            absenceId: {type: integer, format: int64}
            runId: {type: integer, format: int64, nullable: true}
            searched: {type: integer}
            offered: {type: integer}
            settled: {type: integer}
            uncovered: {type: integer}
        absence: {$ref: '#/components/schemas/AbsenceCase'}
    RecordAbsenceRequest:
      type: object
      required: [caregiverId, startDate, endDate]
      properties:
        caregiverId: {type: integer, format: int64}
        type: {type: string, enum: [SICK, ANNUAL, EMERGENCY, OTHER], default: OTHER}
        startDate: {type: string, format: date}
        endDate: {type: string, format: date}
        reason: {type: string, maxLength: 255}
    RequestAbsenceRequest:
      type: object
      required: [startDate, endDate]
      properties:
        type: {type: string, enum: [SICK, ANNUAL, EMERGENCY, OTHER], default: OTHER}
        startDate: {type: string, format: date}
        endDate: {type: string, format: date}
        reason: {type: string, maxLength: 255}
    OwnAbsence:
      type: object
      properties:
        id: {type: integer, format: int64}
        type: {type: string, enum: [SICK, ANNUAL, EMERGENCY, OTHER]}
        startDate: {type: string, format: date}
        endDate: {type: string, format: date}
        reason: {type: string, nullable: true}
        status: {type: string, enum: [PENDING, APPROVED, REJECTED]}
...
  parameters:
...
    AbsenceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
...
    NotFound:
      description: No resource with this id.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    Forbidden:
      description: Authenticated, but not permitted to act on this resource (e.g. an elder outside the caller's scope).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    BadRequest:
      description: The request body or a parameter is malformed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: A business rule refused the request; the detail says which.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: >
        Server-side session established by `POST /auth/login`. State-changing requests must also echo
        the `XSRF-TOKEN` cookie value in an `X-XSRF-TOKEN` header.